8/27/12 End of Night Commit
Game is nearing playability. Will likely finish the core mechanics
tomorrow. Then on to making this my own brand of Roguelike.

• Merged most of XOutlast's code into MapGen, which has become not only
a map generator, but general map handler along with handling the curses
window.
• Changed the start method to be inside the Game class, which should
help for saving/loading multiple different games later on when I
implement it.

+ Added fog of war so that it actually works with objects other than
the player
+ Added enemies(though they don't do much). They take a turn after the
player moves or otherwise does something that requires time.
+ Added fullscreen toggling by pressing the backslash key
+ Game now also exits/ends after pressing the escape key
+ Added in a check to see how many enemies are left which prints a
debug message to console if the player has killed all the enemies, for
later once I add in the message bar